refactor(DTO): Reorganize LegalEntity verbose DTOs
