+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/util/BusinessPartnerVerboseValues.kt
@@ -21,17 +21,15 @@ package org.eclipse.tractusx.bpdm.gate.util
 
 import com.neovisionaries.i18n.CountryCode
 import org.eclipse.tractusx.bpdm.common.dto.*
-import org.eclipse.tractusx.bpdm.common.dto.response.*
+import org.eclipse.tractusx.bpdm.common.dto.response.PhysicalPostalAddressVerboseDto
 import org.eclipse.tractusx.bpdm.common.dto.response.type.TypeKeyNameVerboseDto
 import org.eclipse.tractusx.bpdm.common.model.BusinessStateType
 import org.eclipse.tractusx.bpdm.common.model.ClassificationType
 import org.eclipse.tractusx.bpdm.common.model.DeliveryServiceType
-import org.eclipse.tractusx.bpdm.common.service.toDto
 import org.eclipse.tractusx.bpdm.gate.api.model.*
 import org.eclipse.tractusx.bpdm.gate.api.model.request.BusinessPartnerInputRequest
 import org.eclipse.tractusx.bpdm.gate.api.model.request.BusinessPartnerOutputRequest
 import org.eclipse.tractusx.bpdm.gate.api.model.response.*
-import java.time.Instant
 import java.time.LocalDateTime
 
 object BusinessPartnerVerboseValues {
@@ -78,18 +76,6 @@ object BusinessPartnerVerboseValues {
     val businessStateType1 = BusinessStateType.ACTIVE
     val businessStateType2 = BusinessStateType.INACTIVE
 
-    val legalForm1 = LegalFormDto(
-        technicalKey = "LF1",
-        name = "Limited Liability Company",
-        abbreviation = "LLC",
-    )
-
-    val legalForm2 = LegalFormDto(
-        technicalKey = "LF2",
-        name = "Gemeinschaft mit beschränkter Haftung",
-        abbreviation = "GmbH",
-    )
-
     val bpState1 = BusinessPartnerStateDto(
         validFrom = businessStatusValidFrom1,
         validTo = businessStatusValidUntil1,
@@ -120,30 +106,6 @@ object BusinessPartnerVerboseValues {
         type = identifierTypeTechnicalKey3,
         value = identifierValue3,
         issuingBody = null
-    )
-
-    val classification1 = ClassificationVerboseDto(
-        value = "Sale of motor vehicles",
-        code = "code1",
-        type = ClassificationType.NACE.toDto()
-    )
-
-    val classification2 = ClassificationVerboseDto(
-        value = "Data processing, hosting and related activities",
-        code = "code2",
-        type = ClassificationType.NACE.toDto()
-    )
-
-    val classification3 = ClassificationVerboseDto(
-        value = "Other information service activities",
-        code = "code3",
-        type = ClassificationType.NACE.toDto()
-    )
-
-    val classification4 = ClassificationVerboseDto(
-        value = "Financial and insurance activities",
-        code = "code4",
-        type = ClassificationType.NACE.toDto()
     )
 
 
@@ -243,26 +205,6 @@ object BusinessPartnerVerboseValues {
         legalForm = "LF1",
         states = listOf(legalEntityBusinessStatus1),
         classifications = listOf(bpClassification1Dto, bpClassification2Dto),
-    )
-
-    val leBusinessStatus1 = LegalEntityStateVerboseDto(
-        description = "Active",
-        validFrom = LocalDateTime.of(2020, 1, 1, 0, 0),
-        validTo = LocalDateTime.of(2021, 1, 1, 0, 0),
-        type = TypeKeyNameVerboseDto(
-            technicalKey = BusinessStateType.ACTIVE,
-            name = BusinessStateType.ACTIVE.getTypeName(),
-        )
-    )
-
-    val leBusinessStatus2 = LegalEntityStateVerboseDto(
-        description = "Insolvent",
-        validFrom = LocalDateTime.of(2019, 1, 1, 0, 0),
-        validTo = LocalDateTime.of(2022, 1, 1, 0, 0),
-        type = TypeKeyNameVerboseDto(
-            technicalKey = BusinessStateType.INACTIVE,
-            name = BusinessStateType.INACTIVE.getTypeName(),
-        )
     )
 
     val siteBusinessStatus1 = SiteStateDto(
@@ -407,70 +349,6 @@ object BusinessPartnerVerboseValues {
         floor = "Floor Two",
         door = "Door Two",
         street = StreetDto("TODO", ""),
-    )
-
-    val legalEntityResponsePool1 = PoolLegalEntityVerboseDto(
-        legalName = "Business Partner Name",
-        legalAddress = LogisticAddressVerboseDto(
-            bpna = "BPNA0000000001XY",
-            physicalPostalAddress = address1,
-            bpnLegalEntity = "BPNL0000000000XY",
-            bpnSite = "BPNS0000000001XY",
-            createdAt = Instant.now(),
-            updatedAt = Instant.now()
-        ), legalEntity = LegalEntityVerboseDto(
-            bpnl = "BPNL0000000000XY",
-            legalShortName = "short1",
-            legalForm = legalForm1,
-            states = listOf(leBusinessStatus1),
-            classifications = listOf(classification1, classification2),
-            currentness = Instant.now(),
-            createdAt = Instant.now(),
-            updatedAt = Instant.now(),
-        )
-
-    )
-
-    val legalEntityResponsePool2 = PoolLegalEntityVerboseDto(
-        legalName = "Another Organisation Corp", legalAddress = LogisticAddressVerboseDto(
-            bpna = "BPNA0000000002XY",
-            physicalPostalAddress = address2,
-            bpnLegalEntity = "BPNL0000000001XZ",
-            bpnSite = "BPNS0000000002XY",
-            createdAt = Instant.now(),
-            updatedAt = Instant.now()
-        ), legalEntity = LegalEntityVerboseDto(
-            bpnl = "BPNL0000000001XZ",
-            legalShortName = "short3",
-            legalForm = legalForm2,
-            states = listOf(leBusinessStatus2),
-            classifications = listOf(classification3, classification4),
-            currentness = Instant.now(),
-            createdAt = Instant.now(),
-            updatedAt = Instant.now(),
-        )
-    )
-
-    val legalEntityResponseGate1 = LegalEntityVerboseDto(
-        bpnl = "BPNL0000000000XY",
-        legalShortName = "short1",
-        legalForm = legalForm1,
-        states = listOf(leBusinessStatus1),
-        classifications = listOf(classification1, classification2),
-        currentness = Instant.now(),
-        createdAt = Instant.now(),
-        updatedAt = Instant.now(),
-    )
-
-    val legalEntityResponseGate2 = LegalEntityVerboseDto(
-        bpnl = "BPNL0000000001XZ",
-        legalShortName = "short3",
-        legalForm = legalForm2,
-        states = listOf(leBusinessStatus2),
-        classifications = listOf(classification3, classification4),
-        currentness = Instant.now(),
-        createdAt = Instant.now(),
-        updatedAt = Instant.now(),
     )
 
     //New Values for Logistic Addresses Tests
